requirements.txt
jinja2>=3.1.0
playwright>=1.40.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
